% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bayesmc}
\alias{bayesmc}
\title{Bayesian method for high-dimensional variable selection}
\usage{
bayesmc(Dm, Xmat, b, om1, om2, niter, psample, initsurv, nreport, fitsurv)
}
\arguments{
\item{Dm}{the D matrix}

\item{Xmat}{the design matrix}

\item{b}{the prior distribution parameter for beta, normal std}

\item{om1}{the prior distribution parameter for omega}

\item{om2}{the piror distribution parameter for omega}

\item{niter}{number of iteration}

\item{psample}{the sampling probability for updading regresson coefficient}

\item{initsurv}{initial survival probabilities at end of study}

\item{nreport}{every how many iterations to output parameters}

\item{fitsurv}{the survival parameters optimization function}
}
\description{
Bayesian method for high-dimensional variable selection
}
