% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighDimCR.R
\name{fitsurv}
\alias{fitsurv}
\title{Fit survival function, used for Bayesian simulation}
\usage{
fitsurv(parm, Dm, eta)
}
\arguments{
\item{parm}{the initial parameter value}

\item{Dm}{the D matrix}

\item{eta}{equals to X*bea}
}
\description{
Fit survival function, used for Bayesian simulation
}
