\name{icensmis-package}
\alias{icensmis-package}
\alias{icensmis}
\docType{package}
\title{
	Study Design and Data Analysis in the presence of error-prone diagnostic tests and self-reported outcomes
}
\description{
This package contains two main functions: \code{\link{icpower}} to calculate power and sample size in the presence of error-prone diagnostic tests or self-reported outcomes, \code{\link{icmis}} to analyze such data using a likelihood based approach. When diagnostic tests are perfect, this reduces to interval censored data. The function \code{\link{icpowerpf}} can be used to calculate power and sample size for interval censored data with significantly improved computational efficiency.
	
}
\details{
\tabular{ll}{
Package: \tab icensmis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-10\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Xiangdong Gu and Raji Balasubramanian

Maintainer: Xiangdong Gu <xdgu@schoolph.umass.edu>
}
