% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icmis.R
\name{plot_surv}
\alias{plot_surv}
\title{Plot survival function}
\usage{
plot_surv(obj)
}
\arguments{
\item{obj}{model output object}
}
\description{
This function plots survival function with confidence interval 
from model output
}
